Coupa shahminiaru (#273)
Coupa: 
Create a supplier and invoice in Coupa for each purchase order retrieved from IBM Maximo
Use this template to create a supplier and invoice in Coupa for each purchase order retrieved from IBM Maximo. A Slack message is then sent to the intended channel with the supplier and invoice details.